+++ Desktop.Robot/Desktop.Robot.csproj
@@ -24,11 +24,16 @@
   </ItemGroup>
 
   <ItemGroup>
-	<None Include="osx.os">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-	  <PackagePath>build\</PackagePath>
-	  <Pack>true</Pack>
-	</None>
+    <None Include="osx.os">
+        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+      <PackagePath>build\</PackagePath>
+      <Pack>true</Pack>
+    </None>
+    <None Include="osx_arm.os">
+        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+      <PackagePath>build\</PackagePath>
+      <Pack>true</Pack>
+    </None>
     <None Include="x11.os">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
       <PackagePath>build\</PackagePath>
