+++ Desktop.Robot/Source/osx_arm/Makefile
@@ -1,0 +1,6 @@
+CC=gcc -shared
+FRAMEWORKS:= -framework Foundation -framework Cocoa
+SOURCE=main.mm
+OUT=-o ../../osx_arm.os
+all:
+	$(CC) $(SOURCE) $(FRAMEWORKS) $(OUT)
